[Fix] 수정 내역
1. 도전과제 클리어 안되는 버그 수정(데이터베이스 오타( WALLPAPAER02, 03, 06, FIRDGE06 ))
2. 도전과제 TYPE ID 검증과정 추가(ID에 맞는 데이터가 없을 경우 로그 발생)
3. 손님이 나와있는 중에 스킨 변경시, 나와있는 손님은 스킨 및 능력치가 변경되지 않도록 수정
4. 손님 스킨 능력치 적용(디버깅전)
5. 서버에서 데이터를 불러온 후 평점이 적용되지 않는 문제 수정

+++ Assets/Scripts/FeverSystem/FeverUI/UIFever.cs
@@ -24,7 +24,7 @@ public class UIFever : MonoBehaviour
     private Vector3 _tmpButtonScale;
     private FeverSystem _ferverSystem;
 
-    private int _tmpFeverGauge = 0;
+    private float _tmpFeverGauge = 0;
     public void Init(FeverSystem ferverSystem)
     {
         _ferverSystem = ferverSystem;
